@aisakaki 2023-04-26
> The direction makes sense to me!
> 
> The overall suggestions are:
> 
> * Use the unambiguous names such as original/remap/overwrite paddr/laddr/len
> * Do deep copy to bl/bptr only when necessary, the only case seems to be at the duplicate-for-write time.
> 
> Fixing the readability issues is necessary for the next round of review.

Thanks for review, I will solve these issues soon!