@aisakaki 2023-05-03
> Sorry, my review might be slow due to Ceph deduplication paper work which has been accepted by USENIX ATC'23.

thanks, it doesn't matter~