@cyx1231st 2023-06-26
Due to the variety of the possible combinitions about writing data, a separate PR is needed to extend the UTs with synthetic wirte tests (probably start from `transaction_manager_test_t::test_random_writes_concurrent()`):
* Randomly introduce remap_pin/overwrite_pin.
* Randomly introduce reserved extents.
* Move the synthetic test after the simpler tests.